\name{coef.tsvets.estimate}
\alias{coef}
\alias{coef.tsvets.estimate}
\alias{logLik.tsvets.estimate}
\alias{logLik}
\title{
Estimated Model Extractor Functions
}
\description{
Coefficient, log-likelihood and performance metrics from an estimated tsvets model.
}
\usage{
\method{coef}{tsvets.estimate}(object, ...)
\method{logLik}{tsvets.estimate}(object, ...)
}
\arguments{
\item{object}{
An object of class \dQuote{tsvets.estimate}
}
\item{\dots}{
Not currently used.
}
}
\author{
Alexios Galanos
}