% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tscov.tsvets.estimate}
\alias{tscov.tsvets.estimate}
\alias{tscov}
\alias{tscor}
\alias{tscor.tsvets.estimate}
\title{Model Covariance and Correlation}
\usage{
\method{tscov}{tsvets.estimate}(object, ...)

\method{tscor}{tsvets.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate}.}

\item{...}{not currently used.}
}
\value{
A Matrix object.
}
\description{
Extracts the estimated covariance or correlation matrices.
}
