% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate.tsvets.spec}
\alias{estimate.tsvets.spec}
\alias{estimate}
\title{Model Estimation}
\usage{
\method{estimate}{tsvets.spec}(
  object,
  solver = "nlminb",
  control = list(trace = 0, iter.max = 200, eval.max = 1000),
  autodiff = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.spec}.}

\item{solver}{currently \dQuote{L-BFGS-B} from \dQuote{optim}, \dQuote{nlminb}, 
\dQuote{solnp}, \dQuote{gosolnp} and \dQuote{nloptr} bound constrained solvers 
are supported. For the case of autodiff, only \dQuote{optim} and \dQuote{nlminb} 
are supported, with the latter allowing for the use of the hessian.}

\item{control}{solver control parameters.}

\item{autodiff}{whether to use automatic differentiation for estimation.
This makes use of the tsvetsad package.}

\item{...}{additional arguments to the \dQuote{gosolnp} solver. In the case 
when autodiff is TRUE, then the only other arguments are \dQuote{use_hessian} 
(logical) which works with the nlminb solver and \dQuote{silent} (logical) 
which optionally turns on messages from the TMB package (default is TRUE).}
}
\value{
An object of class \dQuote{tsvets.estimate}
}
\description{
Estimates a model given a specification object using
maximum likelihood.
}
\details{
Minimization of the negative of the log likelihood function of the 
vector Additive ETS model with a soft barrier based on the stability constraint 
of the model.
}
\references{
Athanasopoulos, G and de Silva, A. (2012), 
\emph{Multivariate Exponential Smoothing for Forecasting Tourist Arrivals}, 
Journal of Travel Research 51(5) 640–-652.\cr
de Silva, A., R. Hyndman, and R. D. Snyder. (2010).\emph{The Vector 
Innovations Structural Time Series Framework: A Simple Approach 
to Multivariate Forecasting}, Statistical Modelling (10) 353--74.
}
