% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tsvets.estimate}
\alias{summary.tsvets.estimate}
\alias{summary}
\title{Model Estimation Summary}
\usage{
\method{summary}{tsvets.estimate}(object, weights = NULL, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate}.}

\item{weights}{a vector of weights (summing to 1) which is passed to the 
\code{tsmetrics} method for generating weighted performance metrics.}

\item{...}{not currently used.}
}
\value{
A printout of the parameter summary, model type and some model metrics.
}
\description{
Summary method for class \dQuote{tsvets.estimate}
}
