% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{tsbacktest.tsvets.spec}
\alias{tsbacktest.tsvets.spec}
\alias{tsbacktest}
\title{Walk Forward Model Backtest}
\usage{
\method{tsbacktest}{tsvets.spec}(
  object,
  start = floor(NROW(object$target$y_orig)/2),
  end = NROW(object$target$y_orig),
  h = 1,
  alpha = NULL,
  aggregate = FALSE,
  weights = NULL,
  solver = "nlminb",
  trace = FALSE,
  autodiff = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.spec}.}

\item{start}{numeric data index from which to start the backtest.}

\item{end}{numeric data index on which to end the backtest. The backtest will
end 1 period before that date in order to have at least 1 out of sample value
to compare against.}

\item{h}{forecast horizon. As the expanding window approaches the \dQuote{end},
the horizon will automatically shrink to the number of available out of sample
periods.}

\item{alpha}{optional numeric vector of coverage rates for which to calculate
the quantiles.}

\item{aggregate}{whether to create an aggregate series from the forecasts 
(see \code{\link{tsaggregate}} for details).}

\item{weights}{numeric weights vector for the aggregation.}

\item{solver}{solver to use.}

\item{trace}{whether to show the progress bar. The user is expected to have
set up appropriate handlers for this using the \dQuote{progressr} package.}

\item{autodiff}{whether to use automatic differentiation for estimation.
This makes use of the tsvetsad package.}

\item{...}{additional arguments passed to the \dQuote{auto_clean} function.}
}
\value{
A list with the following data.tables:
\itemize{
\item prediction : the backtest table with forecasts and actuals by series
\item metrics: a summary performance table showing metrics by 
forecast horizon and series
}
}
\description{
Generates an expanding window walk forward backtest.
}
\note{
The function can use parallel functionality as long as the user has
set up a \code{\link[future]{plan}} using the future package.
}
