\name{tscov}
\alias{tscov.tsvets.estimate}
\alias{tscor.tsvets.estimate}
\title{
Model Covariance and Correlation
}
\description{
Extracts the estimated covariance or correlation matrices.
}
\usage{
\method{tscov}{tsvets.estimate}(object, ...)
\method{tscor}{tsvets.estimate}(object, ...)
}
\arguments{
\item{object}{
An object of class \dQuote{tsvets.estimate}.
}
\item{\dots}{
Not currently used.
}
}
\value{
A Matrix object.
}
\author{
Alexios Galanos
}