% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmetrics.tsvets.predict}
\alias{tsmetrics.tsvets.predict}
\alias{tsmetrics}
\alias{tsmetrics.tsvets.estimate}
\title{Performance Metrics}
\usage{
\method{tsmetrics}{tsvets.predict}(object, actual, weights = NULL, alpha = 0.1, ...)

\method{tsmetrics}{tsvets.estimate}(object, weights = NULL, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate} or \dQuote{tsvets.predict}}

\item{actual}{the actual data matched to the dates of the forecasts.}

\item{weights}{a vector of series importance weights (which sum to 1), used 
in generating weighted performance metrics.}

\item{alpha}{the coverage level for distributional forecast metrics.}

\item{...}{not currently used.}
}
\description{
Performance metrics from an estimated or predicted tsvets model.
}
