% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsprofile.R
\name{tsprofile.tsvets.estimate}
\alias{tsprofile.tsvets.estimate}
\alias{tsprofile}
\title{Model Simulation Based Profiling}
\usage{
\method{tsprofile}{tsvets.estimate}(
  object,
  h = 1,
  nsim = 100,
  seed = NULL,
  trace = FALSE,
  solver = "nlminb",
  autodiff = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate}.}

\item{h}{the forecast horizon on which to evaluate performance metrics.}

\item{nsim}{the number of paths to generate.}

\item{seed}{an object specifying if and how the random number generator
should be initialized. See the simulate documentation for more details.}

\item{trace}{whether to show the progress bar. The user is expected to have
set up appropriate handlers for this using the \dQuote{progressr} package.}

\item{solver}{choice of solver to use for the estimation of the paths.}

\item{autodiff}{whether to use automatic differentiation for estimation.
This makes use of the tsvetsad package.}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{tsvets.profile}.
}
\description{
Profiling of model dynamics using simulation/estimation/prediction.
}
\details{
The function profiles an estimated model by simulating and then 
estimating multiple paths from the assumed DGP while leaving h values out for 
prediction evaluation. Each simulated path is equal to the size of the original 
dataset plus h additional values, and initialized with the initial state vector 
from the model. A data.table matrix is returned with the distribution of the 
coefficients from each path estimation as well as a stats table with the MAPE, 
BIAS and MSLRE by horizon, simulation and series.
}
\note{
The function can use parallel functionality as long as the user has set
up a \code{\link[future]{plan}} using the future package.
}
