% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvets-package.R
\docType{package}
\name{tsvets-package}
\alias{tsvets}
\alias{tsvets-package}
\title{tsvets: Vector ETS Models}
\description{
Vector Exponential Smoothing Models.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

Other contributors:
\itemize{
  \item James Nesbit \email{jmcgnesbit@gmail.com} [contributor]
  \item Keith O'Hara \email{kth.ohr@gmail.com} [contributor]
}

}
