% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.tsvets.estimate}
\alias{fitted.tsvets.estimate}
\alias{fitted}
\title{Model Fitted Values}
\usage{
\method{fitted}{tsvets.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate}.}

\item{...}{not currently used.}
}
\description{
Extract the fitted values from an estimated model.
}
