% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.tsvets.estimate}
\alias{residuals.tsvets.estimate}
\alias{residuals}
\title{Model Residuals}
\usage{
\method{residuals}{tsvets.estimate}(object, raw = FALSE, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate}.}

\item{raw}{raw residuals are the model based values in transformed space 
(when either Box Cox or Logistic have been used as transformations).}

\item{...}{not currently used.}
}
\value{
An xts matrix of the model residuals.
}
\description{
Extract the residual values from an estimated model.
}
