% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsdecompose.tsvets.estimate}
\alias{tsdecompose.tsvets.estimate}
\alias{tsdecompose}
\title{Model Decomposition}
\usage{
\method{tsdecompose}{tsvets.estimate}(object, simplify = FALSE, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate}.}

\item{simplify}{simplification of the returned states aggregates the level 
and slope (if present) into a Trend, Seasonal, X and Irregular components.}

\item{...}{not currently used.}
}
\value{
A list of xts matrices with the state components (including error).
}
\description{
Decomposes the estimated model or prediction into its component
parts (states).
}
