% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdiagnose.R
\name{tsdiagnose.tsvets.estimate}
\alias{tsdiagnose.tsvets.estimate}
\alias{tsdiagnose}
\title{Model Diagnostics}
\usage{
\method{tsdiagnose}{tsvets.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsvets.estimate}.}

\item{...}{not currently used.}
}
\value{
A list of tables (printed out and returned invisibly) with diagnostics 
including the eigenvalues of the D matrix as well as Normality and 
multivariate outlier tests.
}
\description{
Creates a short summary of model based diagnostics.
}
